split gl2f.local into modules
